debug : The algorithm for determining forward leaning posture was separated from StrikeNeck.
