+++ FowardLeaningDetector/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace FowardLeaningDetector
+{
+    public class Class1
+    {
+
+    }
+}
