+++ StrikeNeck/StrikeNeck.csproj
@@ -58,14 +58,9 @@
 
 		<PackageReference Include="Microsoft.Data.SqlClient" Version="5.1.1" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0" />
-		<PackageReference Include="Microsoft.ML" Version="2.0.1" />
-		<PackageReference Include="Microsoft.ML.FastTree" Version="2.0.1" />
-		<PackageReference Include="Microsoft.ML.OnnxRuntime" Version="1.15.1" />
-		<PackageReference Include="SixLabors.ImageSharp" Version="3.0.2" />
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Imaging\mlModel\" />
 		<PackageReference Include="SkiaSharp" Version="2.88.5" />
 		<PackageReference Include="Xamarin.Essentials" Version="1.8.0" />
 	</ItemGroup>
